Add files via upload
This is a automatic file sorter code in file explorer. First i have imported os which is the operating system and shutil which enables one to do some high operations.
The files have to be in a folder which have to be  created in the notebook then one can move the file to their specific folders.